praudit: impossible condition
/code/illumos-gate/usr/src/tools/proto/root_i386-nd/opt/onbld/bin/i386/smatch: format.c:769 major_64() warn: impossible condition '(maj > 4294967295) => (0-u32max > u32max)'

+++ usr/src/cmd/praudit/format.c
@@ -762,9 +762,9 @@ minor_64(uint64_t dev)
 static int32_t
 major_64(uint64_t dev)
 {
-	uint32_t maj;
+	uint64_t maj;
 
-	maj = (uint32_t)(dev >> NBITSMINOR64);
+	maj = dev >> NBITSMINOR64;
 
 	if (dev == NODEV || maj > MAXMAJ64) {
 		errno = EINVAL;
